core/app/mAsystent/requirements.txt
absl-py==1.3.0
aiohttp==3.8.6
aiosignal==1.3.1
albumentations==1.3.0
annotated-types==0.6.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
array-record==0.2.0
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
brotlipy==0.7.0
cachetools==5.2.0
certifi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_5d968ni_yn/croot/certifi_1671487774636/work/certifi
cffi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_ab19r4bji3/croot/cffi_1670423206034/work
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cryptography @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_20erp7jjos/croot/cryptography_1665612656373/work
cycler==0.11.0
dataclasses-json==0.6.1
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dm-tree==0.1.8
entrypoints==0.4
etils==1.2.0
exceptiongroup==1.1.3
executing==1.1.1
faiss-cpu==1.7.4
fastjsonschema==2.16.2
flatbuffers==22.10.26
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
frozenlist==1.4.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.59.0
graphviz==0.20.1
grpcio @ file:///Users/builder/miniconda3/envs/prefect/conda-bld/grpcio_1637592348692/work
h11==0.14.0
h5py @ file:///Users/builder/miniconda3/envs/prefect/conda-bld/h5py_1637162753555/work
httpcore==0.18.0
httpx==0.25.0
idna @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_771olrhiqw/croot/idna_1666125579282/work
ImageHash==4.3.1
imageio==2.22.2
importlib-metadata==5.0.0
importlib-resources==5.12.0
imutils==0.5.4
install==1.3.5
ipykernel==6.16.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.21.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
langchain==0.0.316
langsmith==0.0.44
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
marshmallow==3.20.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mediapipe-silicon==0.9.1
mistune==2.0.4
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.7
notebook==6.5.1
notebook_shim==0.2.0
numpy==1.23.4
oauthlib==3.2.2
openai==0.28.1
opencv-contrib-python==4.7.0.72
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
outcome==1.3.0
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.31
protobuf==3.20.0
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.4.2
pydantic_core==2.10.1
pydot==1.4.2
pydotplus==2.0.2
Pygments==2.13.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PyPDF2==3.0.1
pyrsistent==0.18.1
PySocks @ file:///Users/ktietz/Code/oss/ci_pkgs/pysocks_1626781349491/work
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
qudida==0.0.4
regex==2023.10.3
requests @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_b15e28ca-768e-4951-b4c6-8a7476e43183hrjpnkg9/croots/recipe/requests_1657734638595/work
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.2
selenium==4.14.0
Send2Trash==1.8.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.22
stack-data==0.5.1
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-datasets==4.9.2
tensorflow-estimator==2.10.0
tensorflow-macos==2.10.0
tensorflow-metadata==1.13.1
tensorflow-metal==0.6.0
termcolor==2.0.1
terminado==0.17.0
threadpoolctl==3.1.0
tifffile==2022.10.10
tiktoken==0.5.1
tinycss2==1.2.1
toml==0.10.2
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
torchviz==0.0.2
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
trio==0.22.2
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3 @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_e11ioy9zkf/croot/urllib3_1670526989300/work
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.2
zipp==3.10.0
